Create leetcode_1297.js
gaaah learnt that the method to get the number of keys in a map is not `Object.keys(map).length()` but is just a direct data property accessor via `map.size`. Overcomplicated the problem logic.
At least the `string.slice(left,right)` functionality is useful -> just need to remember that `.slice()` calls are exclusive, not inclusive.