Create leetcode_1604.js
Learnt that usage data for generating alerts will not always arrive sequentially, even in a given day -> the records can arrive in a more "random" fashion, and we may not be able to treat the data as a real-time stream. The upstream collections of all usage data may render prerequisite to their future downstream processing.