Create leetcode_89.cpp
Why is it that using the std::bitset<n> template library from C++ is harder? Seriously?
And why can't this stinkin template utility not accept a non-constant, variable template parameter? Gaaah I have to declare additional memory for problems which didn't even need it in the first place ( e.g. std::bitset<16> for a case of n = 2). This is more expensive .