Create leetcode_636.js
Practicing working on solving for the exclusive time of functions - called recursively or non recursively - on a function call stack. 
The word problem emulates real-world situations as is good to solve, since it's challenging to handle the cases of the next function call under execution if it is terminating or not.