Create triangleType.rust
So the problem would've been easy in python, but you're trying to pick up RUST, which is why you're having to spend 20 minutes to even do this problem. If it wasn't for the time you have to spend doing the following
a. learning a new language's standard library features
b. learning what NOT to do in another language
c. learning how to work with mutable types versus non-mutable types
... well, life may have been easier